splits asset_kitchen .hpp files into load.hpp and convert.hpp to better hide the heavy work
