+++ include/Lib/asset_kitchen/assets.hpp
@@ -1,11 +1,13 @@
 #pragma once
 
-#include "Lib/core/core.hpp"
-#include "Lib/opengl/.hpp"
-#include "Lib/render/.hpp"
-#include "Lib/scene/.hpp"
-
 #include "descriptions.hpp"
+
+#include "Lib/core/core.hpp"
+#include "Lib/core/named.hpp"
+#include "Lib/opengl/texture.hpp"
+#include "Lib/render/mesh.hpp"
+#include "Lib/render/material.hpp"
+#include "Lib/scene/core.hpp"
 
 struct Assets
 {
@@ -42,41 +44,7 @@ struct Assets
 			load_gltf(name);
 	}
 
-	void load_glsl_program(Name const & name)
-	{
-		auto const loaded_data = GLSL::Program::Load(descriptions.glsl.get(name));
-
-		if (auto expected = GLSL::Program::Convert(loaded_data))
-		{
-			programs.get_or_generate(name) = expected.into_result();
-			program_errors.get_or_generate(name) = {};
-		}
-		else
-		{
-			programs.get_or_generate(name) = {};
-			program_errors.get_or_generate(name) = expected.into_error();
-		}
-	}
-
-	void load_glsl_uniform_block(Name const & name)
-	{
-		auto const loaded_data = GLSL::UniformBlock::Load(descriptions.uniform_block.get(name));
-
-		if (auto expected = GLSL::UniformBlock::Convert(loaded_data))
-		{
-			uniform_blocks.get_or_generate(name) = expected.into_result();
-			program_errors.get_or_generate(name) = {};
-		}
-		else
-		{
-			uniform_blocks.get_or_generate(name) = {};
-			program_errors.get_or_generate(name) = expected.into_error();
-		}
-	}
-
-	void load_gltf(Name const & name)
-	{
-		auto const gltf_data = GLTF::Load(descriptions.gltf.get(name));
-		GLTF::Convert(gltf_data, textures, materials, primitives, meshes, scene_tree);
-	}
+	void load_glsl_program(Name const & name);
+	void load_glsl_uniform_block(Name const & name);
+	void load_gltf(Name const & name);
 };
