+++ include/Lib/asset_kitchen/glsl/program/load.hpp
@@ -1,12 +1,8 @@
 #pragma once
-#pragma message("----Read ASSET/PROGRAM/.Hpp----")
+#pragma message("----Read ASSET/PROGRAM/load.Hpp----")
 
 #include "Lib/core/core.hpp"
-#include "Lib/core/expected.hpp"
-#include "Lib/file_management/core.hpp"
-#include "Lib/file_management/json_utils.hpp"
-#include "Lib/opengl/glsl.hpp"
-#include "Lib/opengl/shader.hpp"
+#include "Lib/opengl/core.hpp"
 
 namespace GLSL::Program
 {
@@ -36,8 +32,4 @@ namespace GLSL::Program
 	};
 
 	LoadedData Load(Description const & description);
-
-	Expected<GL::ShaderProgram, std::string> Convert(LoadedData const & loaded);
-
-	std::pair<Name, Description> Parse(File::JSON::JSONObj o, std::filesystem::path const & root_dir);
 }
