+++ include/Lib/asset_kitchen/glsl/uniform_block/convert.hpp
@@ -1,5 +1,7 @@
 #pragma once
-#pragma message("----Read ASSET/BLOCK/.Hpp----")
+#pragma message("----Read ASSET/BLOCK/convert.Hpp----")
+
+#include "load.hpp"
 
 #include "Lib/core/expected.hpp"
 #include "Lib/file_management/core.hpp"
@@ -8,17 +10,6 @@
 
 namespace GLSL::UniformBlock
 {
-	struct LoadedData
-	{
-		std::string source;
-	};
-
-	struct Description
-	{
-		std::filesystem::path path;
-	};
-
-	LoadedData Load(Description const & description);
 	Expected<GL::UniformBlock, std::string> Convert(LoadedData const & loaded);
 	std::pair<Name, Description> Parse(File::JSON::JSONObj o, std::filesystem::path const & root_dir);
 }
