+++ include/Lib/asset_kitchen/gltf/load.hpp
@@ -1,14 +1,8 @@
 #pragma once
-#pragma message("----Read ASSET/GLTF/.Hpp----")
+#pragma message("----Read ASSET/GLTF/load.Hpp----")
 
 #include "Lib/core/core.hpp"
 #include "Lib/core/named.hpp"
-#include "Lib/file_management/core.hpp"
-#include "Lib/file_management/json_utils.hpp"
-#include "Lib/opengl/texture.hpp"
-#include "Lib/render/material.hpp"
-#include "Lib/render/mesh.hpp"
-#include "Lib/scene/core.hpp"
 
 namespace GLTF
 {
@@ -161,15 +155,4 @@ namespace GLTF
 	};
 
 	LoadedData Load(Description const & description);
-
-	void Convert(
-		LoadedData const & loaded,
-		Managed<GL::Texture2D> & textures,
-		Managed<unique_one<Render::IMaterial>> & materials,
-		Managed<Geometry::Primitive> & primitives,
-		Managed<Render::Mesh> & meshes,
-		::Scene::Tree & scene_tree
-	);
-
-	std::pair<Name, Description> Parse(File::JSON::JSONObj o, std::filesystem::path const & root_dir);
 }
